feat(dashboard): open the request feature form in a drawer
Replace the /request-feature page with a drawer opened by the sidebar
button, since the form is too small to justify a full page. Removes the
whitelabel route, the page, and the two column section.
