+++ apps/dashboard/widgets/WhitelabelShell.tsx
@@ -10,7 +10,8 @@ import { WhitelabelConnectWallet } from "@/shared/components/wallet/WhitelabelCo
 import daoConfigByDaoId from "@/shared/dao-config";
 import type { DaoIdEnum } from "@/shared/types/daos";
 import { cn } from "@/shared/utils/cn";
-import { getDaoPagePath, WHITELABEL_ROUTES } from "@/shared/utils/whitelabel";
+import { RequestFeatureDrawer } from "@/features/request-feature/RequestFeatureDrawer";
+import { getDaoPagePath } from "@/shared/utils/whitelabel";
 import { WHITELABEL_NAV_ITEMS } from "@/shared/utils/whitelabelNav";
 import { WhitelabelHeader } from "@/widgets/WhitelabelHeader";
 import { WhitelabelHeaderMobile } from "@/widgets/WhitelabelHeaderMobile";
@@ -26,13 +27,9 @@ export const WhitelabelShell = ({
   const pathname = usePathname();
   const [isSidebarCollapsed, setIsSidebarCollapsed] = useState(false);
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
+  const [isRequestFeatureOpen, setIsRequestFeatureOpen] = useState(false);
   const daoConfig = daoConfigByDaoId[daoId];
   const externalRequestFeatureLink = daoConfig.whitelabel?.requestFeatureLink;
-  const requestFeatureHref = getDaoPagePath({
-    daoId,
-    pathname,
-    page: WHITELABEL_ROUTES.requestFeature,
-  });
 
   const navItems = useMemo(
     () =>
@@ -53,6 +50,13 @@ export const WhitelabelShell = ({
         daoId={daoId}
         isCollapsed={isSidebarCollapsed}
         onToggleCollapse={() => setIsSidebarCollapsed((current) => !current)}
+        onRequestFeatureClick={() => setIsRequestFeatureOpen(true)}
+      />
+
+      <RequestFeatureDrawer
+        daoId={daoId}
+        isOpen={isRequestFeatureOpen}
+        onClose={() => setIsRequestFeatureOpen(false)}
       />
 
       <div className="flex min-w-0 flex-1 flex-col">
@@ -101,12 +105,12 @@ export const WhitelabelShell = ({
               })}
             </nav>
 
-            <Button
-              variant="outline"
-              asChild
-              className="mt-4 w-full justify-center"
-            >
-              {externalRequestFeatureLink ? (
+            {externalRequestFeatureLink ? (
+              <Button
+                variant="outline"
+                asChild
+                className="mt-4 w-full justify-center"
+              >
                 <a
                   href={externalRequestFeatureLink}
                   target="_blank"
@@ -118,19 +122,23 @@ export const WhitelabelShell = ({
                 >
                   Request feature
                 </a>
-              ) : (
-                <Link
-                  href={requestFeatureHref}
-                  onClick={() => setIsMobileMenuOpen(false)}
-                  data-umami-event="feature_request_click"
-                  data-umami-event-source="whitelabel_shell"
-                  data-ph-event="feature_request_click"
-                  data-ph-source="whitelabel_shell"
-                >
-                  Request feature
-                </Link>
-              )}
-            </Button>
+              </Button>
+            ) : (
+              <Button
+                variant="outline"
+                className="mt-4 w-full justify-center"
+                onClick={() => {
+                  setIsMobileMenuOpen(false);
+                  setIsRequestFeatureOpen(true);
+                }}
+                data-umami-event="feature_request_click"
+                data-umami-event-source="whitelabel_shell"
+                data-ph-event="feature_request_click"
+                data-ph-source="whitelabel_shell"
+              >
+                Request feature
+              </Button>
+            )}
           </div>
         </div>
